refactor: 자동차 객체에 우승자 판별 책임 위임
