+++ src/main/kotlin/Application.kt
@@ -2,8 +2,7 @@ fun main() {
     val inputReader = InputReader()
     val outputPrinter = OutputPrinter()
 
-    val judge = Judge()
-    val raceOrganizer = RaceOrganizer(inputReader, outputPrinter, judge)
+    val raceOrganizer = RaceOrganizer(inputReader, outputPrinter)
 
     raceOrganizer.hostRace()
 }
